pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyquickref"
version = "0.1.0"
description = "A comprehensive reference for Python functionality with runnable examples"
readme = "README.md"
authors = [
    {name = "PyQuickRef Team", email = "example@example.com"}
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
pyquickref = "pyquickref.cli:main"

[dependency-groups]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=6.1.0",
    "ruff>=0.1.6",
    "ty>=0.0.1",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
]

[tool.setuptools.packages.find]
include = ["pyquickref*"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "D", "UP", "ANN", "B", "A", "C4", "SIM", "ERA"]
ignore = ["D203", "D213", "ANN401", "COM812"]

[tool.ruff.lint.per-file-ignores]
"pyquickref/examples/functional.py" = ["E731"]
"pyquickref/examples/strings.py" = ["UP032", "ERA001"]
"pyquickref/examples/design_patterns.py" = ["E731"]
"pyquickref/examples/stdlib_tools.py" = ["E731"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.environment]
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v"

[tool.coverage.run]
source = ["pyquickref"]
relative_files = true
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
